ランダム移動のテコ入れ（停止時間に応じて）

+++ work/Resist_multi.java
@@ -438,7 +438,7 @@ public class Resist_multi {
         dif_col = r.pld_col - r.col;
         dif_row = r.pld_row - r.row;
 
-        int enhanced_eps = Constants.epsiron * r.not_move_count;
+        double enhanced_eps = Constants.epsiron * r.not_move_count;
         if(enhanced_eps == 0) enhanced_eps = Constants.epsiron;
 
         // RECONFIG!!!!!!
@@ -874,7 +874,7 @@ public class Resist_multi {
         int next_col = r.col;
         int next_row = r.row;
 
-        int enhanced_eps = Constants.epsiron * r.not_move_count;
+        double enhanced_eps = Constants.epsiron * r.not_move_count;
         if(enhanced_eps == 0) enhanced_eps = Constants.epsiron;
 
         // I will add the pattern when these two values are "=" later.
